adjustments about organization list
