+++ locales/en.json
@@ -20,7 +20,7 @@
   "navbar-link-dark-mode": "Dark/Light mode",
   "navbar-user-profile": "User Profile",
   "navbar-org-profile": "Org Profile",
-  "navbar-organization": "Org. list",
+  "navbar-organization-list": "Org. list",
   "navbar-organization-profile": "Organization Profile",
   "body-home-section01-title-text": "A space for exchanging",
   "body-home-section01-title-carousel": "capacity, knowledge, skills, services",
