--- src/app/organization/page.tsx
@@ -1,10 +1,0 @@
-import BaseWrapper from "@/components/BaseWrapper";
-import FeedPage from "./components/page";
-
-export default function Organization(){
-
-    return (
-        <BaseWrapper>
-            <FeedPage/>
-        </BaseWrapper>
-    )}
