+++ src/app/organization_list/page.tsx
@@ -1,0 +1,10 @@
+import BaseWrapper from "@/components/BaseWrapper";
+import OrganizationList from "./OrganizationList";
+
+export default function OrganizationListWrapper(){
+
+    return (
+        <BaseWrapper>
+            <OrganizationList/>
+        </BaseWrapper>
+    )}
